refactor(246): update tanstack-query plugin architecture
